+++ src/warning.ts
@@ -1,0 +1,9 @@
+type IWarning = [boolean, string];
+
+export function warning(warnings: IWarning[]): void {
+  warnings.forEach(([isThrow, message]) => {
+    if (isThrow) {
+      throw new Error(message);
+    }
+  });
+}