+++ tests/index.test.ts
@@ -1,3 +1,0 @@
-test('hello test', () => {
-  expect(1).toEqual(1);
-});