+++ src/warning.ts
@@ -1,4 +1,4 @@
-type IWarning = [boolean, string];
+export type IWarning = [boolean, string];
 
 export function warning(warnings: IWarning[]): void {
   warnings.forEach(([isThrow, message]) => {