+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Dev
 
+# 0.3.0 (May 11, 2020)
+
 1. Added store config test
 2. Added pice config test
 3. Added pice state test